NF: move TagsList to its own file
