+++ AnkiDroid/src/test/java/com/ichi2/anki/dialogs/tags/TagsListTest.java
@@ -15,8 +15,6 @@
  */
 package com.ichi2.anki.dialogs.tags;
 
-import com.ichi2.anki.dialogs.tags.TagsDialog.TagsList;
-
 import org.junit.Before;
 import org.junit.Test;
